backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.7
Pillow==10.1.0
reportlab==4.0.7
zipfile36==0.1.3
aiofiles==23.2.1
pydantic==2.5.0
httpx==0.25.2
replicate==0.22.0
huggingface-hub==0.19.4
transformers==4.36.0
torch==2.1.1
torchvision==0.16.1
diffusers==0.24.0
accelerate==0.25.0 